backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings>=2.10.1
pydantic[email]
email-validator

# AI — Gemini SDK (legacy analyze endpoint)
google-generativeai==0.8.3

# AI Agent (LangGraph 0.3.x ecosystem)
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.37,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langgraph>=0.2.0,<1.0.0
langsmith
duckduckgo-search

# Database
supabase==2.9.0
psycopg2-binary==2.9.9

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
pdfplumber==0.10.3

# Payments
stripe==11.1.0

# Utils
aiofiles==24.1.0
httpx==0.27.0
